+++ lib/presentation/app.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../application/user/state/delete_user_result.dart';
-import '../util/routing/router.dart';
+import '../config/router.dart';
 import 'component/async_value_handler.dart';
 import 'component/dialog.dart';
 import 'component/loading.dart';
